# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "东软顾养中心管理系统API文档",
        "title": "东软顾养中心 API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/api",
    "paths": {
...
        "/services": {
            "get": {
                "description": "分页获取服务项目定义列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "服务管理"
                ],
                "summary": "获取服务项目列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "类别",
                        "name": "category",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "状态",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "跳过数量",
                        "name": "skip",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "每页数量",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "description": "添加新的服务项目定义",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "服务管理"
                ],
                "summary": "创建服务项目",
                "parameters": [
                    {
                        "description": "服务项目信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.Service"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "创建成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/services/customer-service/{id}/end": {
            "put": {
                "description": "根据购买记录ID手动结束一项服务",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "服务管理"
                ],
                "summary": "结束客户服务",
                "parameters": [
                    {
                        "type": "string",
                        "description": "购买记录ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "结束信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.EndServiceRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "结束成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/services/customer/{customer_id}": {
            "get": {
                "description": "获取指定客户已购买的服务列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "服务管理"
                ],
                "summary": "获取客户服务列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "客户ID",
                        "name": "customer_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/services/purchase": {
            "post": {
                "description": "为客户购买指定的服务项目",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "服务管理"
                ],
                "summary": "客户购买服务",
                "parameters": [
                    {
                        "description": "购买信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.PurchaseServiceRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "购买成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/services/{id}": {
            "get": {
                "description": "根据ID获取服务项目详情",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "服务管理"
                ],
                "summary": "获取详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "服务项目ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "description": "根据ID更新服务项目信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "服务管理"
                ],
                "summary": "更新服务项目",
                "parameters": [
                    {
                        "type": "string",
                        "description": "服务项目ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "服务项目信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.Service"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "description": "根据ID删除服务项目定义",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "服务管理"
                ],
                "summary": "删除服务项目",
                "parameters": [
                    {
                        "type": "string",
                        "description": "服务项目ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.EndServiceRequest": {
            "type": "object",
            "required": [
                "end_date"
            ],
            "properties": {
                "end_date": {
                    "type": "string"
                }
            }
        },
...
        "domain.PurchaseServiceRequest": {
            "type": "object",
            "required": [
                "customer_id",
                "service_id",
                "start_date"
            ],
            "properties": {
                "customer_id": {
                    "type": "string"
                },
                "service_id": {
                    "type": "string"
                },
                "start_date": {
                    "type": "string"
                }
            }
        },
...
        "domain.Service": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "category": {
                    "description": "服务类别：\"医疗\"/\"生活\"/\"娱乐\"等",
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "description": "服务名称",
                    "type": "string"
                },
                "price": {
                    "description": "服务价格",
                    "type": "number"
                },
                "status": {
                    "description": "状态：\"启用\"/\"停用\"",
                    "type": "string"
                },
                "unit": {
                    "description": "计价单位：\"次\"/\"月\"/\"年\"",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}